Split NativeEditorApp into app-shell, scene, and overlay files
NativeEditorApp.kt had grown to 1,427 lines mixing three unrelated
concerns. Separate them into same-package files so each reads on its own
and the editor is easier to extend:

- NativeEditorApp.kt (419) keeps the Compose app shell: main (so the
  mainClass entry point is unchanged), the window/shortcut wiring, file
  dialogs, the NativeEditorApp composable, the canvas factory, and the
  resource-locator bootstrap.
- EditorScene.kt (929) holds the Ardor3D Scene/Updater document: scene
  mutation, save/open/import, picking, and the render/update loop.
- EditorOverlays.kt (124) holds the stateless overlay-geometry factories
  (grid, starter cube, light gizmo), which never touched instance state,
  as internal top-level functions.

Pure code-move: same declarations and bodies, same package, so no call
sites change. The only semantic difference is the three overlay factories
going from private methods to internal top-level functions. Compiles with
no warnings or unused imports; the editor test suite stays green.
